[DEV-1757] move fee & payment
