+++ src/subdomains/core/buy-crypto/routes/crypto-route/crypto-route.controller.ts
@@ -8,7 +8,6 @@ import { JwtPayload } from 'src/shared/auth/jwt-payload.interface';
 import { RoleGuard } from 'src/shared/auth/role.guard';
 import { UserRole } from 'src/shared/auth/user-role.enum';
 import { AssetDtoMapper } from 'src/shared/models/asset/dto/asset-dto.mapper';
-import { TransactionHelper } from 'src/shared/payment/services/transaction-helper';
 import { PaymentInfoService } from 'src/shared/services/payment-info.service';
 import { Util } from 'src/shared/utils/util';
 import { BuyCryptoService } from 'src/subdomains/core/buy-crypto/process/services/buy-crypto.service';
@@ -16,6 +15,7 @@ import { HistoryDto } from 'src/subdomains/core/history/dto/history.dto';
 import { FeeDirectionType } from 'src/subdomains/generic/user/models/user/user.entity';
 import { UserService } from 'src/subdomains/generic/user/models/user/user.service';
 import { DepositDtoMapper } from 'src/subdomains/supporting/address-pool/deposit/dto/deposit-dto.mapper';
+import { TransactionHelper } from 'src/subdomains/supporting/payment/services/transaction-helper';
 import { CryptoRoute } from './crypto-route.entity';
 import { CryptoRouteService } from './crypto-route.service';
 import { CreateCryptoRouteDto } from './dto/create-crypto-route.dto';
