+++ src/subdomains/core/core.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
+import { PaymentModule } from '../supporting/payment/payment.module';
 import { BuyCryptoModule } from './buy-crypto/buy-crypto.module';
-import { FeeModule } from './fee/fee.module';
 import { HistoryModule } from './history/history.module';
 import { LiquidityManagementModule } from './liquidity-management/liquidity-management.module';
 import { MonitoringModule } from './monitoring/monitoring.module';
@@ -17,7 +17,7 @@ import { StatisticModule } from './statistic/statistic.module';
     SellCryptoModule,
     StatisticModule,
     LiquidityManagementModule,
-    FeeModule,
+    PaymentModule,
   ],
   controllers: [],
   providers: [],
