+++ src/subdomains/core/sell-crypto/route/dto/sell.dto.ts
@@ -1,8 +1,8 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { Blockchain } from 'src/integration/blockchain/shared/enums/blockchain.enum';
 import { FiatDto } from 'src/shared/models/fiat/dto/fiat.dto';
-import { MinAmount } from '../../../../../shared/payment/dto/min-amount.dto';
 import { DepositDto } from '../../../../supporting/address-pool/deposit/dto/deposit.dto';
+import { MinAmount } from '../../../../supporting/payment/dto/min-amount.dto';
 
 export class SellDto {
   @ApiProperty()
