+++ src/subdomains/core/sell-crypto/sell-crypto.module.ts
@@ -1,7 +1,6 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { BlockchainModule } from 'src/integration/blockchain/blockchain.module';
-import { PaymentModule } from 'src/shared/payment/payment.module';
 import { SharedModule } from 'src/shared/shared.module';
 import { UserModule } from 'src/subdomains/generic/user/user.module';
 import { AddressPoolModule } from 'src/subdomains/supporting/address-pool/address-pool.module';
@@ -10,6 +9,7 @@ import { BankModule } from 'src/subdomains/supporting/bank/bank.module';
 import { FiatOutputModule } from 'src/subdomains/supporting/fiat-output/fiat-output.module';
 import { NotificationModule } from 'src/subdomains/supporting/notification/notification.module';
 import { PayInModule } from 'src/subdomains/supporting/payin/payin.module';
+import { PaymentModule } from 'src/subdomains/supporting/payment/payment.module';
 import { BuyCryptoModule } from '../buy-crypto/buy-crypto.module';
 import { BuyFiatNotificationService } from './process/buy-fiat-notification.service';
 import { BuyFiatRegistrationService } from './process/buy-fiat-registration.service';
