+++ src/subdomains/supporting/payin/payin.module.ts
@@ -2,9 +2,9 @@ import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { BlockchainModule } from 'src/integration/blockchain/blockchain.module';
 import { ChainalysisModule } from 'src/integration/chainalysis/chainalysis.module';
-import { PaymentModule } from 'src/shared/payment/payment.module';
 import { SharedModule } from 'src/shared/shared.module';
 import { SellCryptoModule } from 'src/subdomains/core/sell-crypto/sell-crypto.module';
+import { PaymentModule } from 'src/subdomains/supporting/payment/payment.module';
 import { DexModule } from '../dex/dex.module';
 import { NotificationModule } from '../notification/notification.module';
 import { PayoutModule } from '../payout/payout.module';
