+++ src/subdomains/supporting/payment/controllers/fee.controller.ts
@@ -3,9 +3,9 @@ import { AuthGuard } from '@nestjs/passport';
 import { ApiBearerAuth, ApiExcludeEndpoint, ApiTags } from '@nestjs/swagger';
 import { RoleGuard } from 'src/shared/auth/role.guard';
 import { UserRole } from 'src/shared/auth/user-role.enum';
-import { CreateFeeDto } from './dto/create-fee.dto';
-import { Fee } from './fee.entity';
-import { FeeService } from './fee.service';
+import { CreateFeeDto } from '../dto/create-fee.dto';
+import { Fee } from '../entities/fee.entity';
+import { FeeService } from '../services/fee.service';
 
 @ApiTags('Fee')
 @Controller('fee')
