+++ src/subdomains/supporting/payment/dto/min-amount.dto.ts
