+++ src/subdomains/supporting/payment/dto/tx-spec.dto.ts
