+++ src/subdomains/supporting/payment/entities/transaction-specification.entity.ts
