+++ src/subdomains/supporting/payment/exceptions/fee-limit-exceeded.exception.ts
