+++ src/subdomains/supporting/payment/repositories/transaction-specification.repository.ts
